HDS-464 using directories
