+++ src/Middleware/integrations/ordercloud.integrations.library/blobs/BlobService.cs
@@ -31,7 +31,7 @@ namespace ordercloud.integrations.library
         Task<bool> CreateContainerAsync(string containerName, bool isPublic);
         Task<List<CloudBlobContainer>> ListContainers();
         Task CopyBlobs();
-        Task TransferBlobs(string sourceContainer, string destinationContainer, string blobName);
+        Task TransferBlobs(string sourceContainer, string destinationContainer, string blobName, string directoryName = "");
 
     }
     public class OrderCloudIntegrationsBlobService : IOrderCloudIntegrationsBlobService
@@ -134,44 +134,24 @@ namespace ordercloud.integrations.library
             return results;
         }
 
-        public async Task TransferBlobs(string sourceContainer, string destinationContainer, string blobName)
+        public async Task TransferBlobs(string sourceContainer, string destinationContainer, string blobName, string directoryName = "")
         {
-            string directoryName = "webfolder";
+            // Download all files from blob storage to this folder
+            // Upload all these files in this folder to $web in blob storage
             await this.Init();
-            CreateDirectory(directoryName);
-            await DownloadBlob(sourceContainer, blobName);
-            await UploadBlob(destinationContainer, blobName);
-            DeleteDirectory(directoryName);
+            await DownloadBlob(sourceContainer, blobName, directoryName);
+            await UploadBlob(destinationContainer, blobName, directoryName);
         }
         
-        private void CreateDirectory(string directoryName)
-        {
-            if (!Directory.Exists(directoryName))
-            {
-                Directory.CreateDirectory(directoryName);
-            }
-        }
 
-        private void DeleteDirectory(string directoryName)
-        {
-            if (!Directory.Exists(directoryName))
-            {
-                Directory.Delete(directoryName);
-            }
-        }
-
-        private async Task DownloadBlob(string sourceContainer, string blobName)
+        private async Task DownloadBlob(string sourceContainer, string blobName, string directory)
         {
             try
             {
                 CloudBlobContainer sourceBlobContainer = Client.GetContainerReference(sourceContainer);
                 ICloudBlob sourceBlob = await sourceBlobContainer.GetBlobReferenceFromServerAsync(blobName);
-                if(!Directory.Exists(_webfolder))
-                {
-                    Directory.CreateDirectory(_webfolder);
-                }
 
-                await sourceBlob.DownloadToFileAsync(_webfolder + "/" + blobName.Replace("/", "_"), System.IO.FileMode.Create);
+                await sourceBlob.DownloadToFileAsync(directory + "/" + blobName.Replace("/", "_"), System.IO.FileMode.Create);
 
                 //modifying the app configs.
                 var configTest = "{'hostedApp': true, 'appname': 'headstartDemo' }";
@@ -184,7 +164,7 @@ namespace ordercloud.integrations.library
             
         }
 
-        private async Task UploadBlob(string destinationContainer, string blobName)
+        private async Task UploadBlob(string destinationContainer, string blobName, string directory)
         {
             string filepath = "https://headstartdemo.blob.core.windows.net/buyerweb";
             CloudBlobContainer destBlobContainer = Client.GetContainerReference(destinationContainer);
@@ -195,7 +175,7 @@ namespace ordercloud.integrations.library
             {
                 destBlob.Properties.ContentType = contentType;
             }
-            await destBlob.UploadFromFileAsync(_webfolder + "/" + blobName.Replace("/", "_"));
+            await destBlob.UploadFromFileAsync(directory + "/" + blobName.Replace("/", "_"));
         }
 
         private string GetContentType(string fileName)
